My solution
During the researching process, I have discovered that variable "$this->_form_id" can refer both to an array and to an integer in the same time, depending on where it is used. In order to counter that, and prevent issues with implode function, I have created a "$formId" variable as an error handler. Please refer to a code.